requirements.txt
numpy
captcha
tensorflow
pandas
scikit-learn
pillow 
numpy